edited layout a bit
edited layout and changed scripts a bit
