+++ GhostClub/Assets/Art Assets/Skyboxes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 57e31cbe533004c4b8712a7fdd0bddab
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
