+++ GhostClub/Assets/Art Assets/Skyboxes/Skybox.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ca7b778cd5ff4452caa17c02daef8168
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
