+++ GhostClub/Assets/Scripts/HostBattleHUD.cs
@@ -5,15 +5,13 @@ using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class BattleHUD : MonoBehaviour
+public class HostBattleHUD : MonoBehaviour
 {
     public TextMeshProUGUI nameText;
     public TextMeshProUGUI physicalDamageText;
-    public TextMeshProUGUI magicalDamageText;
     public TextMeshProUGUI physicalHPtext;
-    public TextMeshProUGUI magicalHPtext;
+    
 
-    public GameObject bossObj;
     public Unit currentUnit;
 
     public void Start()
@@ -21,23 +19,22 @@ public class BattleHUD : MonoBehaviour
         //NS: this part needs work...how to find the specific instantiated Object?
         //VC: I added tags for the Boss, Host and Summon prefabs,
         //    and wired them up in the GameManager
-        currentUnit = GameManager.instance.boss.GetComponent<Unit>();
+        currentUnit = GameManager.instance.host.GetComponent<Unit>();
     }
     
     public void Update()
     {
-        SetupHUD(currentUnit);
+        SetupHostHUD(currentUnit);
     }
 
-    public void SetupHUD(Unit unit)
+    public void SetupHostHUD(Unit unit)
     {
         unit = currentUnit;
         
         nameText.text = "Name: " + unit.unitName;
         physicalDamageText.text = "Phys DMG: " + unit.physicalDamage;
-        magicalDamageText.text = "MP DMG: " + unit.magicalDamage;
         physicalHPtext.text = "Phys HP: " + unit.currentPhysicalHP;
-        magicalHPtext.text = "MP HP: " +  unit.currentMagicalHP;
+        
     }
