+++ GhostClub/Assets/Scripts/HostBattleHUD.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 71895cd8fa1c94caa8aadcc956328370
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 100
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
