+++ GhostClub/Assets/Scripts/SummonBattleHUD.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b0f23cc10f1504dcc91107ed37b754b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 100
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
